+++ src/XlsxCell.h
@@ -36,19 +36,21 @@ class XlsxCell {
 
 public:
 
-  XlsxCell(rapidxml::xml_node<>* cell): cell_(cell) {
+  // if possible, provide guess at row and column based on position within xml
+  XlsxCell(rapidxml::xml_node<>* cell, int row = -1, int col = -1): cell_(cell) {
     rapidxml::xml_attribute<>* ref = cell_->first_attribute("r");
-    if (ref == NULL)
-      Rcpp::stop("Invalid cell: lacks ref attribute");
-
-    location_ = parseRef(ref->value());
+    if (ref == NULL) {
+      location_ = std::make_pair(row, col);
+    } else {
+      location_ = parseRef(ref->value());
+    }
   }
 
-  int row() {
+  int row() const {
     return location_.first;
   }
 
-  int col() {
+  int col()  const {
     return location_.second;
   }