+++ NEWS.md
@@ -1,5 +1,11 @@
 # readxl 0.1.1.9000
 
+* Improved parsing of sheet geometry for xlsx. (#240, @jennybc).
+
+    - Better handling of leading and embedded blank rows and explicit row skipping. (#224, #194, #178, #156, #101)
+    - Worksheets that are completely empty or that contain only column names no longer error, but return a tibble with zero rows. (#222, #144, #65)
+    - Location is inferred for cells that do not declare their location (e.g. xlsx written by JMP). (#163, #102)
+
 * Logic for sheet lookup in xlsx is more robust. Improves compatibility with xlsx written by tools other than Excel and/or xlsx containing chartsheets. (#233, #104, #200, #168, #116, @jimhester, @jennybc)
 
 * Support multiple NA values, e.g., `read_excel("missing-values.xls", na = c("NA", "1"))` (#13, #56, @jmarshallnz).